requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
configparser==7.2.0
DateTime==5.5
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
libpcap==1.11.0b25
logging==0.4.9.6
MarkupSafe==3.0.2
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
pkg_about==1.3.7
plotly==6.2.0
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pydeck==0.9.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scapy==2.6.1
setuptools==80.9.0
shutils==0.1.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.42
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing==3.7.4.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
zope.interface==7.2
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
python-dotenv==1.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1

